+++ pkg/protocol/rest/server.go
@@ -91,8 +91,8 @@ func RunServer(config *myconfig.ServerConfig) error {
 			return err
 		}
 	}
-	scanningService := service.NewScanningService(config)
-	if err := scanningService.TestEngine(); err != nil {
+	apiService := service.NewApiService(config)
+	if err := apiService.TestEngine(); err != nil {
 		zlog.S.Warnf("Scanning engine test failed. Scan requests are likely to fail.")
 		zlog.S.Warnf("Please make sure that %v is accessible", config.Scanning.ScanBinary)
 	}
@@ -103,9 +103,10 @@ func RunServer(config *myconfig.ServerConfig) error {
 	router.HandleFunc("/api/health", service.HealthCheck).Methods(http.MethodGet)
 	router.HandleFunc("/api/health-check", service.HealthCheck).Methods(http.MethodGet)
 	router.HandleFunc("/api/metrics/{type}", service.MetricsHandler).Methods(http.MethodGet)
-	router.HandleFunc("/api/scan/direct", scanningService.ScanDirect).Methods(http.MethodPost)
-	router.HandleFunc("/api/file_contents/{md5}", scanningService.FileContents).Methods(http.MethodGet)
-	//router.HandleFunc("/api/sbom/attribution", scanningService.FileContents).Methods(http.MethodPost)
+	router.HandleFunc("/api/file_contents/{md5}", apiService.FileContents).Methods(http.MethodGet)
+	router.HandleFunc("/api/license/obligations/{license}", apiService.LicenseDetails).Methods(http.MethodGet)
+	router.HandleFunc("/api/scan/direct", apiService.ScanDirect).Methods(http.MethodPost)
+	router.HandleFunc("/api/sbom/attribution", apiService.SbomAttribution).Methods(http.MethodPost)
 	srv := &http.Server{
 		Handler: router,
 		Addr:    fmt.Sprintf("%s:%s", config.App.Addr, config.App.Port),