+++ CHANGELOG.md
@@ -9,6 +9,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Upcoming changes...
 
+## [0.5.0] - 2023-01-07
+### Added
+- Added attribution endpoint
+- Added license details endpoint
+
+## [0.4.1] - 2023-01-04
+### Added
+- Added API health
+- Added API metrics
+### Fixed
+- Fixed issue with scanning form parameters
 
 ## [0.0.1] - 2022-12-15
 ### Added
@@ -17,3 +28,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added threaded scanning workers
 
 [0.0.1]: https://github.com/scanoss/wayuu2/compare/v0.0.0...v0.0.1
+[0.4.1]: https://github.com/scanoss/wayuu2/compare/v0.0.1...v0.4.1
+[0.5.0]: https://github.com/scanoss/wayuu2/compare/v0.4.1...v0.5.0